Mark 'Faster Spark Application Development and Debugging' as draft
Changes made in this commit mark the 'Faster Spark Application Development and Debugging' article as a draft, hence it won't be live on the website. The author, Vusal Dadalov, and last update date are also added under metadata section, offering essential information about the documentation authorship and its latest update.